Port more example test code to java
