--- org.eclipse.xtext.xtext.ui.examples/projects/homeautomation/org.eclipse.xtext.example.homeautomation.tests/xtend-gen/org/eclipse/xtext/example/homeautomation/tests/AbstractTest.java
@@ -1,100 +1,0 @@
-/**
- * Copyright (c) 2015, 2020 itemis AG (http://www.itemis.eu) and others.
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- * 
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.eclipse.xtext.example.homeautomation.tests;
-
-import org.eclipse.xtend2.lib.StringConcatenation;
-import org.eclipse.xtext.testing.InjectWith;
-import org.eclipse.xtext.testing.XtextRunner;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-@RunWith(XtextRunner.class)
-@InjectWith(RuleEngineInjectorProvider.class)
-@SuppressWarnings("all")
-public abstract class AbstractTest {
-  protected abstract void test(final CharSequence document) throws Exception;
-  
-  @Test
-  public void test1() throws Exception {
-    StringConcatenation _builder = new StringConcatenation();
-    _builder.append("Device Window can be open, closed");
-    _builder.newLine();
-    _builder.append("Device Heater can be on, off");
-    _builder.newLine();
-    _builder.append("Rule \"Save energy\" when Window.open then");
-    _builder.newLine();
-    _builder.append("\t");
-    _builder.append("fire(Heater.off)");
-    _builder.newLine();
-    _builder.append("\t");
-    _builder.append("println(\'Another penny to the piggy bank!\')");
-    _builder.newLine();
-    this.test(_builder);
-  }
-  
-  @Test
-  public void test2() throws Exception {
-    StringConcatenation _builder = new StringConcatenation();
-    _builder.append("Device Window can be open, closed");
-    _builder.newLine();
-    _builder.append("Device Heater can be on, off, error");
-    _builder.newLine();
-    _builder.append("Rule \"Save energy\" when Window.open then");
-    _builder.newLine();
-    _builder.append("\t");
-    _builder.append("for (v : Heater.values)");
-    _builder.newLine();
-    _builder.append("\t\t");
-    _builder.append("val output = \'Heater state \' + v.name");
-    _builder.newLine();
-    _builder.append("\t\t");
-    _builder.append("println(output)");
-    _builder.newLine();
-    _builder.append("\t");
-    _builder.append("fire(Heater.off)");
-    _builder.newLine();
-    this.test(_builder);
-  }
-  
-  @Test
-  public void test3() throws Exception {
-    StringConcatenation _builder = new StringConcatenation();
-    _builder.append("Device Window can be open, closed");
-    _builder.newLine();
-    _builder.append("Device Heater can be on, off");
-    _builder.newLine();
-    _builder.append("Rule \"Save energy\" when Window.open then");
-    _builder.newLine();
-    _builder.append("\t");
-    _builder.append("switch new java.util.Random().nextInt()");
-    _builder.newLine();
-    _builder.append("\t\t");
-    _builder.append("case 1,");
-    _builder.newLine();
-    _builder.append("\t\t");
-    _builder.append("case 2:");
-    _builder.newLine();
-    _builder.append("\t\t\t");
-    _builder.append("fire(Heater.off)");
-    _builder.newLine();
-    _builder.append("\t\t\t");
-    _builder.append("println(\'Another penny to the piggy bank!\')");
-    _builder.newLine();
-    _builder.append("\t\t");
-    _builder.append("default:");
-    _builder.newLine();
-    _builder.append("\t\t\t");
-    _builder.append("fire(Heater.on)");
-    _builder.newLine();
-    _builder.append("\t\t\t");
-    _builder.append("println(\'I do support global warming!\')");
-    _builder.newLine();
-    this.test(_builder);
-  }
-}
